getting fitness values

+++ suricata/src/pesquisa.h
@@ -1,4 +1,4 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-void logFitness(void *, int, int);
+void logFitness(void *, int, float);
